backend/requirements.txt
sentence-transformers
scikit-learn
pandas
spacy
pdfminer.six
pyresparser
transformers
nltk
openai
Flask==2.3.2
python-docx==0.8.11
PyMuPDF==1.22.5
Werkzeug==2.3.6